+++ app/lib/fs-utility.js
@@ -57,7 +57,6 @@ module.exports = {
   },
 
   setDirIcon: function(balloonDir) {
-console.log(getIcon());
     return nodeIcon.setFolderIcon(balloonDir, getIcon(), {
       cmd: getFileiconPath()
     });